prevent double-counting of TIB for empty stage val/times
Cmetric_stages is set to 1 by SOpowerHistogram. The way the code was initially structured, it would ignore the stage information for logicals, thus double-counting the TIB across all the stage. So I updated the code not to double count.
SOpowerHistogram deals with empty stage vals and times so I don't think we need to make them required inputs.